bridge/requirements.txt
pyserial==3.5
python-telegram-bot==13.4.1
requests==2.25.1
